Feat: Use Context API
Following changes were done to utilize React's Context API,
- A new file store/project-management-context.jsx was added to hold the context store.
- State manipulator functions from the App.jsx were copied to the above file.
- The main component in the App.jsx was wrapped with the Context provider component.

Note: The functionality of the project is still running with the help of prop drilling of the state and callback functions.
Next objective will be to use the state from the context API directly in the componetns and remove all drilled props.